Part4 done without 4.6, 4.7 and 4.23

+++ controllers/blogsRoutes.js
@@ -1,8 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+                .find({})
+                .populate('user', {username : 1, name : 1})
   response.json(blogs)
 })
 
@@ -30,21 +34,26 @@ blogsRouter.get('/:id', async (request, response, next) => {
     */
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', middleware.userExtractor,  async (request, response, next) => {
   const body = request.body
+
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
+    user: user._id
   })
 
   
   const savedBlog = await blog.save()
-  response.status(200).json(savedBlog)
  
-
+ 
+  user.blogs = user.blogs.concat(savedBlog.id)
+  await user.save()
+  response.status(201).json(savedBlog)
   /*
   blog.save()
     .then(savedBlog => {
@@ -54,10 +63,18 @@ blogsRouter.post('/', async (request, response, next) => {
     */
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
 
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+ 
+  
+  if(user.id.toString() === blog.user.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    response.status(404).json({ error: 'user not authorized' })
+  }
 
   /** before async/await
    * Blog.findByIdAndRemove(request.params.id)
@@ -70,24 +87,21 @@ blogsRouter.delete('/:id', async (request, response) => {
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-  const body = request.body
+  const blog = request.body
 
-  
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
+  if(!request.token) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
-  /*
-  await Blog.findByIdAndUpdate(request.params.id, blog)
-  response.json(blog)
-  */
-  
-  const updatedBlog  = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+  const updatedBlog = await Blog
+    .findByIdAndUpdate(
+      request.params.id,
+      blog,
+      { new: true, runValidators: true, context: 'query' }
+    )
+
   response.status(200).json(updatedBlog)
   
-    
+
 })
 
 module.exports = blogsRouter
